+++ tools/tesseratos/src/tesseratos/entity_inspector/plugin.cpp
@@ -1,6 +1,7 @@
 #include <imgui.h>
 
 #include <cubos/core/ecs/entity/entity.hpp>
+#include <cubos/core/ecs/name.hpp>
 #include <cubos/core/reflection/reflect.hpp>
 
 #include <cubos/engine/imgui/data_inspector.hpp>
@@ -10,7 +11,7 @@
 #include <tesseratos/entity_selector/plugin.hpp>
 #include <tesseratos/toolbox/plugin.hpp>
 
-using cubos::core::ecs::Entity;
+using cubos::core::ecs::Name;
 using cubos::core::ecs::World;
 using cubos::core::memory::AnyValue;
 using cubos::core::reflection::reflect;
@@ -18,6 +19,73 @@ using cubos::core::reflection::Type;
 
 using cubos::engine::Cubos;
 using cubos::engine::DataInspector;
+using cubos::engine::Entity;
+using cubos::engine::Opt;
+using cubos::engine::Query;
+
+namespace
+{
+    struct State
+    {
+        const Type* relationType;
+    };
+} // namespace
+
+static void addRelationButton(State& state, World& world, Entity entity, bool incoming,
+                              Query<Entity, Opt<const Name&>>& query)
+{
+    std::string suffix = incoming ? "##incoming" : "##outgoing";
+
+    if (ImGui::Button(("Add Relation" + suffix).c_str()))
+    {
+        ImGui::OpenPopup(("Select Relation Type" + suffix).c_str());
+    }
+
+    bool openPopup = false;
+    if (ImGui::BeginPopup(("Select Relation Type" + suffix).c_str()))
+    {
+        for (auto [type, name] : world.types().relations())
+        {
+            if (ImGui::Button(name.c_str()))
+            {
+                state.relationType = type;
+                openPopup = true;
+                ImGui::CloseCurrentPopup();
+            }
+        }
+
+        ImGui::EndPopup();
+    }
+
+    if (openPopup)
+    {
+        ImGui::OpenPopup(("Select Relation Target" + suffix).c_str());
+    }
+
+    if (ImGui::BeginPopup(("Select Relation Target" + suffix).c_str()))
+    {
+        for (auto [target, name] : query)
+        {
+            std::string targetStr =
+                name ? name->value : (std::to_string(target.index) + "#" + std::to_string(target.generation));
+            if (ImGui::Button(targetStr.c_str()))
+            {
+                auto value = AnyValue::defaultConstruct(*state.relationType);
+                if (incoming)
+                {
+                    world.relate(target, entity, value.type(), value.get());
+                }
+                else
+                {
+                    world.relate(entity, target, value.type(), value.get());
+                }
+                ImGui::CloseCurrentPopup();
+            }
+        }
+
+        ImGui::EndPopup();
+    }
+}
 
 void tesseratos::entityInspectorPlugin(Cubos& cubos)
 {
@@ -25,13 +93,30 @@ void tesseratos::entityInspectorPlugin(Cubos& cubos)
     cubos.depends(entitySelectorPlugin);
     cubos.depends(toolboxPlugin);
 
+    cubos.resource<State>();
+
     cubos.system("show Entity Inspector UI")
         .tagged(cubos::engine::imguiTag)
-        .call([](World& world, Toolbox& toolbox, const EntitySelector& entitySelector, DataInspector& dataInspector) {
+        .call([](State& state, World& world, Toolbox& toolbox, const EntitySelector& entitySelector,
+                 DataInspector& dataInspector, Query<Entity, Opt<const Name&>> query) {
             if (!toolbox.isOpen("Entity Inspector"))
             {
                 return;
             }
+
+            auto getName = [&](Entity ent) -> std::string {
+                if (!world.isAlive(ent))
+                {
+                    return "Invalid";
+                }
+
+                if (world.components(ent).has<Name>())
+                {
+                    return world.components(ent).get<Name>().value;
+                }
+
+                return std::to_string(ent.index) + "#" + std::to_string(ent.generation);
+            };
 
             ImGui::Begin("Entity Inspector");
             if (!ImGui::IsWindowCollapsed())
@@ -39,8 +124,8 @@ void tesseratos::entityInspectorPlugin(Cubos& cubos)
                 auto entity = entitySelector.selection;
                 if (!entity.isNull() && world.isAlive(entity))
                 {
-                    ImGui::Text("Entity %d#%d selected", entity.index, entity.generation);
-                    ImGui::Separator();
+                    ImGui::Text("Entity %s selected", getName(entity).c_str());
+                    ImGui::SeparatorText("Components");
 
                     if (ImGui::Button("Add Component"))
                     {
@@ -78,6 +163,54 @@ void tesseratos::entityInspectorPlugin(Cubos& cubos)
                     if (removed != nullptr)
                     {
                         world.components(entity).remove(*removed);
+                        removed = nullptr;
+                    }
+
+                    ImGui::SeparatorText("Incoming Relations");
+                    addRelationButton(state, world, entity, true, query);
+
+                    Entity removedEnt{};
+                    for (auto [type, value, fromEntity] : world.relationsTo(entity))
+                    {
+                        std::string relName = getName(fromEntity) + "@" + type->shortName();
+                        if (ImGui::CollapsingHeader(relName.c_str()))
+                        {
+                            dataInspector.edit(*type, value);
+                            if (ImGui::Button("Remove Relation"))
+                            {
+                                removed = type;
+                                removedEnt = fromEntity;
+                            }
+                        }
+                    }
+
+                    if (removed != nullptr)
+                    {
+                        world.unrelate(removedEnt, entity, *removed);
+                        removed = nullptr;
+                    }
+
+                    ImGui::SeparatorText("Outgoing Relations");
+                    addRelationButton(state, world, entity, false, query);
+
+                    for (auto [type, value, toEntity] : world.relationsFrom(entity))
+                    {
+                        std::string name = type->shortName() + "@" + getName(toEntity);
+                        if (ImGui::CollapsingHeader(name.c_str()))
+                        {
+                            dataInspector.edit(*type, value);
+                            if (ImGui::Button("Remove Relation"))
+                            {
+                                removed = type;
+                                removedEnt = toEntity;
+                            }
+                        }
+                    }
+
+                    if (removed != nullptr)
+                    {
+                        world.unrelate(entity, removedEnt, *removed);
+                        removed = nullptr;
                     }
                 }
                 else