+++ CHANGELOG.md
@@ -39,6 +39,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Possibility of building the core library as a shared library (#1052, **@RiscadoA**).
 - Possibility of building the engine library as a shared library (#1053, **@RiscadoA**).
 - Plugins system argument, which can be used to dynamically add and remove plugins (#1034, **@RiscadoA**).
+- Show relations on the Entity Inspector (#907, **@RiscadoA**).
+- Allow adding and removing relations through the Entity Inspector (#1022, **@RiscadoA**).
 
 ### Changed